Finish developing and test ROI functions for Dataset
Rename:
data_proc_common.py -> roi_utils.py

A new function roi_utils.calc_dataset_ROIs() performs ROI calculation
and (optionally) merging of the new ROI dimensions with an old ROI
dimension for xarray.DataSet.

The test for roi_utils.calc_dataset_ROIs() is:
/test/test_calc_dataset_ROIs_and_combine.py

If a Dataset variable does not contain some of the coordinates that should
be collapsed into ROIs, then ROI calculations is performed on the existing
coordinates only, with the appropriate renaming of the new ROI dimension and
coordinate values by this dimension.
Example:
	Qxy variable contains x and y coordinates
	Qx variable contains x coordinate
	If we collapse x and y coordinates into ROIs, then in the new Dataset:
	Qxy will contain xyROI_num, xyROI_name, and xyROI_name2 coordinates
	Qx will contain xROI_num, xROI_name, and xROI_name2 coordinates
	Values of xyROI_numalues of xyROI_num, xROI_num: 0, 1, 2, ...
	Values of xyROI_name: xyROI0, xyROI1, ...
	Values of xROI_name: xROI0, xROI1, ...
	Values of xyROI_name2:
		ROI_(x=x11-x12)_(y=y11-y12), ROI_(x=x21-x22)_(y=y21-y22), ...
	Values of xROI_name2:
		ROI_(x=x11-x12), ROI_(x=x21-x22), ...

The same logic is used for the "old" ROI dimensions, which should be joined
with the new one.
Example:
	Consider that we collapsed the coords x and y into ROIs for a Dataset with
	the variables Qxyz (with coords x, y, z) and Qxz (with coords x, z).
	Now we want to collapse z into ROIs and join the new zROI_num dimension
	with the old xyROI_num dimension.
	For the variable Qxyz, the dimensions xyROI_num and zROI_num will be joined
	into a dimension xyzROI.
	For the variable Qxz, the dimensions xROI_num and zROI_num will be joined
	into a dimension xzROI.

To implement the aforementioned logic, the following changes were made to
roi_utils.calc_dataset_ROIs() and roi_utils.calc_xarray_ROIs():
	- The name of the new ROI dimension is generated automatically from the names
	of the used coordinates. To do this, set ROIset_dim_name=None.
	- The name of the old ROI dimension is given not explicitly, but with the list
	of the coordinate names that were used for this ROI calculation. For example,
	instead of ROIset_dim_to_combine_var='xyROI', one should set
	ROIset_dim_to_combine_var=['x', 'y'].