Create a class for managing xarray attributes
Class DataAttributes is defined in /data_attributes.py
It is used to accumulate information about steps of a processing pipeline
in the attributes of DataArray / Dataset.

This class allows to:
- Unflatten the 'attr' dictionary from the input DataArray / Dataset
- Add a meaningfule description of a processing step to it
- Flatten the result to put it to the 'attr' of the output DataArray / Dataset

Flattenning is required because DataArray / Dataset is going to be saved into
netcdf format, which does not support nested dictionaries.

The test for DataAttributes class is:
/test/test_data_attr.py