Implement DataFileGroup class for distributed storage of Datasets
DataFileGroup is going to be the main structure for data storage in this
project. It is organized as a pandas table containing paths to the files
with actual data.
Several columns of the table could be specified as 'outer' dimensions of
the data. Thus, each table row (and so each file) corresponds to a specific
point in the space of outer dimensions.
Each file contains an xarray.Dataset object with its 'inner' dimensions.
Inner dimensions are the same for all files.

A sequence of procesing steps that resulted in a DataFileGroup object is
given by DataProcTree class (that replaced an old DataAttributes class).
In the current implementation, all steps belong to a single branch. In the
future, several merging branches will be implemented.
A processing step description contains info about:
	- Its name and the function used to perform it
	- Description of the data produced by this step (variables, inner and
	outer dimensions)
	- Parameters of the processing
	A sequence of procesing steps is stored in each Dataset object in its
	'attrs'	field.

Code for converting old time-frequncy data into the new format is given in
reformat_data_files.py. It gives appropriate names to the variables of
Datasets and puts info about the processing steps into the 'attrs' fields.

Changes in files:
	- data_file_group.py
	  A new file containing DataFileGroup and DataProcTree class definitions
	- data_attributes.py
	  Deleted. DataAttributes class eloved into DataProcTree class and was
	  moved to data_file_group.py
	- test/test_data_attr.py
	  Changed: DataProcTree is used instead of DataAttributes
	- useful.py
	  The functions for flattening / unflattenning a dict were moved here
	  from data_attributes.py
	- reformat_data_files.py
	  A new file with the code that converts old time-frequency data into the
	  DataFileGroup format