Selection of cell pairs that have significant PLV with the same channel
- select_cell_pairs_by_spPLV.py
	select_cell_pairs_by_spPLV()
		Input: DataTable containing PLV's between channels and cells
		(calculated in several time-frequency ROI's), as well as the
		corresponding p-values.
		Output: DataTable containing pairs of cells which have
		significant PLV's with the same channel in a given ROI.

In the given implementation, the algorithm walks through the channels
and, for each channel, it finds all cells that have significant PLV
with this channel in a given time-frequency ROI. PLV is considered as
significant if the corresponding p-value is smaller than a given threshold.
Cells with the firing rates higher than a given max. value are discarded.
For a channel, all found cells are randomly grouped into pairs, and the
information about these pairs is stored into the output DataTable object.

- test/test_dfg_find_trial_pairs
	Test script for select_cell_pairs_by_spPLV()

- workflow6.py
	Selection of the cell pairs is added to the workflow.